# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/person": {
      "delete": {
        "operationId": "PersonController_deletePerson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletePersonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delete a person record including its employee and user record will be used for testing only"
          }
        },
        "summary": "Delete a person",
        "tags": [
          "Person"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "DeletePersonDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}